refactor luggage all Luggage et test

+++ flycheckin/boarding/src/main/java/ca/ulaval/glo4002/flycheckin/boarding/domain/luggage/CheckedLuggage.java
@@ -3,10 +3,10 @@ package ca.ulaval.glo4002.flycheckin.boarding.domain.luggage;
 public class CheckedLuggage extends Luggage {
 
   private static final String TYPE = "checked";
-  private static final double DIMENSION_LIMIT = 158;
+  private static final int DIMENSION_LIMIT = 158;
   private static final double SURPLUS_PRICE = 50;
 
-  public CheckedLuggage(double linearDimension, double weight) {
+  public CheckedLuggage(int linearDimension, int weight) {
     super(linearDimension, weight);
   }
